refactor(broker): split order_locker.rs into directory layout (C32 of PR-3)
Pure mechanical move. order_locker.rs (1583 lines) becomes:

- order_locker/mod.rs (~35 lines): module decls, pub use service::OrderLocker,
  pub use types::RpcRetryConfig, pub(crate) use types::{OrderCommitmentMeta,
  OrderLockerConfig}, and #[cfg(test)] pub(crate) use service::tests so
  shared/prioritization.rs keeps reaching crate::order_locker::tests::*.
- order_locker/error.rs (~55 lines): OrderLockerErr enum + coded_error_impl!.
- order_locker/types.rs (~46 lines): OrderCommitmentMeta (pub(crate)),
  CapacityResult (pub(super) — its fields too, since it's constructed and
  consumed across modules now), OrderLockerConfig (pub(crate)), RpcRetryConfig
  (pub).
- order_locker/service.rs (~1523 lines): OrderLocker<P> struct + impl,
  BrokerService impl, the original test module, and the free helper
  estimate_proving_time_no_load.

External paths crate::order_locker::{OrderLocker, RpcRetryConfig,
OrderCommitmentMeta, OrderLockerConfig} preserved via re-exports. Test path
crate::order_locker::tests::setup_oc_test_context preserved via the cfg-test
re-export.
